backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.3.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.1
Deprecated==1.2.13
Django==3.2.13
django-nose==1.4.7
django-redis==5.2.0
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
isort==5.10.1
kombu==5.2.4
mypy-extensions==0.4.3
nose==1.3.7
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.50
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
